+++ Sources/PulseUI/Features/Inspector/NetworkDetailsView.swift
@@ -30,7 +30,7 @@ struct NetworkDetailsView: View {
 
     @ViewBuilder
     private var contents: some View {
-        if let viewModel = viewModel?.text, !viewModel.isEmpty {
+        if let viewModel = viewModel?.text, viewModel.originalText.length > 0 {
             RichTextView(viewModel: viewModel)
         } else {
             PlaceholderView(imageName: "nosign", title: "Empty")